Move WorkflowController
